Add safer version of EdDSA signing API.
The old API takes the public key as an input.  If the wrong public key
is passed, then the private key may be revealed to the attacker (via the
usual two-signatures-one-nonce vulnerability).  As a countermeasure, the
API now instead rederives the pubkey and asserts equality.  This makes it
twice as slow.

The new API stores the public and private keypair in one struct, which is
initialized by expanding the private key.  This protects against mistakes.

The old API is soft-deprecated.  I might add a deprecation attribute
later.